(Chore) Split the "manage contacts" feature spec up
This spec had become too big and unwieldy; split it up into 3 smaller feature
specs.
